changed acquisition col name
